# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Time tracker and invoice generator API
 * OpenAPI spec version: 0.1.0
 */
export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus];

export const TaskStatus = {
  active: "active",
  completed: "completed",
  archived: "archived",
} as const;

export interface Task {
  id: string;
  title: string;
  description: string;
  status: TaskStatus;
  createdAt: string;
}

export type TaskWithStatsStatus =
  (typeof TaskWithStatsStatus)[keyof typeof TaskWithStatsStatus];

export const TaskWithStatsStatus = {
  active: "active",
  completed: "completed",
  archived: "archived",
} as const;

export interface TaskWithStats {
  id: string;
  title: string;
  description: string;
  status: TaskWithStatsStatus;
  createdAt: string;
  totalSeconds: number;
  unbilledSeconds: number;
  entryCount: number;
}

export interface CreateTaskBody {
  /** @minLength 1 */
  title: string;
  description: string;
}

export type UpdateTaskBodyStatus =
  (typeof UpdateTaskBodyStatus)[keyof typeof UpdateTaskBodyStatus];

export const UpdateTaskBodyStatus = {
  active: "active",
  completed: "completed",
  archived: "archived",
} as const;

export interface UpdateTaskBody {
  /** @minLength 1 */
  title?: string;
  description?: string;
  status?: UpdateTaskBodyStatus;
}

export interface TimeEntry {
  id: string;
  taskId: string;
  taskTitle?: string;
  description: string;
  startedAt: string;
  endedAt?: string | null;
  durationSeconds: number;
  isRunning: boolean;
  invoiceId?: string | null;
}

export interface StartTimerBody {
  /** @minLength 1 */
  description: string;
}

export interface UpdateEntryBody {
  /** @minLength 1 */
  description?: string;
}

export interface ActiveEntryResponse {
  entry?: TimeEntry | null;
}

export interface SummaryResponse {
  clientName: string;
  hourlyRate: number;
  totalTasks: number;
  totalSeconds: number;
  unbilledSeconds: number;
  unbilledAmount: number;
  paidAmount: number;
  outstandingAmount: number;
  recentEntries: TimeEntry[];
  activeEntry: TimeEntry | null;
}

export interface InvoiceLineItem {
  id: string;
  taskId: string;
  taskTitle: string;
  description: string;
  startedAt: string;
  endedAt: string;
  durationSeconds: number;
  amount: number;
}

export type InvoiceStatus = (typeof InvoiceStatus)[keyof typeof InvoiceStatus];

export const InvoiceStatus = {
  unpaid: "unpaid",
  paid: "paid",
} as const;

export interface Invoice {
  id: string;
  invoiceNumber: string;
  clientName: string;
  hourlyRate: number;
  status: InvoiceStatus;
  createdAt: string;
  paidAt?: string | null;
  notes?: string;
  totalSeconds: number;
  totalAmount: number;
  lineItems: InvoiceLineItem[];
}

export type InvoiceSummaryStatus =
  (typeof InvoiceSummaryStatus)[keyof typeof InvoiceSummaryStatus];

export const InvoiceSummaryStatus = {
  unpaid: "unpaid",
  paid: "paid",
} as const;

export interface InvoiceSummary {
  id: string;
  invoiceNumber: string;
  clientName: string;
  status: InvoiceSummaryStatus;
  createdAt: string;
  paidAt?: string | null;
  totalSeconds: number;
  totalAmount: number;
  lineItemCount: number;
}

export interface InvoicePreview {
  clientName: string;
  hourlyRate: number;
  totalSeconds: number;
  totalAmount: number;
  lineItems: InvoiceLineItem[];
}

export interface CreateInvoiceBody {
  notes?: string;
}

export type UpdateInvoiceStatusBodyStatus =
  (typeof UpdateInvoiceStatusBodyStatus)[keyof typeof UpdateInvoiceStatusBodyStatus];

export const UpdateInvoiceStatusBodyStatus = {
  unpaid: "unpaid",
  paid: "paid",
} as const;

export interface UpdateInvoiceStatusBody {
  status: UpdateInvoiceStatusBodyStatus;
}