# lib/api-zod/src/generated/types/createInvoiceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Time tracker and invoice generator API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateInvoiceBody {
  notes?: string;
}