# lib/api-zod/src/generated/types/invoice.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Time tracker and invoice generator API
 * OpenAPI spec version: 0.1.0
 */
import type { InvoiceLineItem } from "./invoiceLineItem";
import type { InvoiceStatus } from "./invoiceStatus";

export interface Invoice {
  id: string;
  invoiceNumber: string;
  clientName: string;
  hourlyRate: number;
  status: InvoiceStatus;
  createdAt: Date;
  paidAt?: Date | null;
  notes?: string;
  totalSeconds: number;
  totalAmount: number;
  lineItems: InvoiceLineItem[];
}