# lib/api-zod/src/generated/types/taskWithStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Time tracker and invoice generator API
 * OpenAPI spec version: 0.1.0
 */
import type { TaskWithStatsStatus } from "./taskWithStatsStatus";

export interface TaskWithStats {
  id: string;
  title: string;
  description: string;
  status: TaskWithStatsStatus;
  createdAt: Date;
  totalSeconds: number;
  unbilledSeconds: number;
  entryCount: number;
}