# lib/api-zod/src/generated/types/updateTaskBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Time tracker and invoice generator API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateTaskBodyStatus } from "./updateTaskBodyStatus";

export interface UpdateTaskBody {
  /** @minLength 1 */
  title?: string;
  description?: string;
  status?: UpdateTaskBodyStatus;
}